+++ src/test/java/it/gov/pagopa/pu/debtpositions/util/UtilitiesTest.java
@@ -1,5 +1,6 @@
 package it.gov.pagopa.pu.debtpositions.util;
 
+import org.apache.commons.lang3.tuple.Pair;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -302,4 +303,28 @@ public class UtilitiesTest {
     );
   }
 
+  @Test
+  void givenDptoIbanWhenResolveIbanAndPostalIbanThenReturnDptoIbans() {
+    String dptoIban = "IT0000000000000000000000000";
+    String dptoPostalIban = "IT0000000000000000000000001";
+    String orgIban = "IT0000000000000000000000002";
+    String orgPostalIban = "IT0000000000000000000000003";
+
+    Pair<String, String> result = Utilities.resolveIbanAndPostalIban(dptoIban, dptoPostalIban, orgIban, orgPostalIban);
+
+    assertEquals(dptoIban, result.getLeft());
+    assertEquals(dptoPostalIban, result.getRight());
+  }
+
+  @Test
+  void givenNullDptoIbanWhenResolveIbanAndPostalIbanThenReturnOrgIbans() {
+    String dptoPostalIban = "IT0000000000000000000000001";
+    String orgIban = "IT0000000000000000000000002";
+    String orgPostalIban = "IT0000000000000000000000003";
+
+    Pair<String, String> result = Utilities.resolveIbanAndPostalIban(null, dptoPostalIban, orgIban, orgPostalIban);
+
+    assertEquals(orgIban, result.getLeft());
+    assertEquals(orgPostalIban, result.getRight());
+  }
 }
